% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Eq_Modulus.R
\name{calculate_Eq_Modulus}
\alias{calculate_Eq_Modulus}
\title{calculate_Eq_Modulus}
\usage{
calculate_Eq_Modulus(Se1, Se2, Se3, Indentation_area, Strains)
}
\arguments{
\item{Se1}{forces at equilibrium at the end of the 1st step}

\item{Se2}{forces at equilibrium at the end of the 2nd step}

\item{Se3}{forces at equilibrium at the end of the 3rd step}

\item{Indentation_area}{the indentation area (based on the radius of the indenter)}

\item{Strains}{list of the strains at the 3 steps}
}
\value{
the calculated equilibrium elastic Modulus
}
\description{
calculate Equilibrium elastic Modulus using the forces at equilibrium from the 3 stress-relaxation steps
}
