% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Ins_Modulus.R
\name{calculate_Ins_Modulus}
\alias{calculate_Ins_Modulus}
\title{calculate_Ins_Modulus}
\usage{
calculate_Ins_Modulus(Sd1, Sd2, Sd3, Indentation_area, Strains)
}
\arguments{
\item{Sd1}{the difference between the peak force and the force just prior the peak at the beginning of the 1st step}

\item{Sd2}{the difference between the peak force and the force just prior the peak at the beginning of the 2nd step}

\item{Sd3}{the difference between the peak force and the force just prior the peak at the beginning of the 3rd step}

\item{Indentation_area}{the indentation area (based on the radius of the indenter)}

\item{Strains}{list of the strains at the 3 steps}
}
\value{
the calculated equilibrium elastic Modulus
}
\description{
calculate Instantaneous elastic Modulus using the forces at equilibrium from the 3 stress-relaxation steps
}
