% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_moduli_multifiles2.R
\name{calculate_moduli_multifiles2}
\alias{calculate_moduli_multifiles2}
\title{calculate_moduli_multifiles2}
\usage{
calculate_moduli_multifiles2(
  folder,
  Thicknesses,
  Strain,
  indenterradius,
  poisson_eq = 0.5,
  poisson_inst = 0.5,
  smfl = 19,
  EqDataPoints = 201
)
}
\arguments{
\item{folder}{path to a folder with the MACH-1 text files}

\item{Thicknesses}{a list with the thicknesses for of the sample for the three steps}

\item{Strain}{a list with the strains for of the sample for the three steps}

\item{indenterradius}{the radius of the indeter}

\item{poisson_eq}{the poisson ratio for the equilibrium modulus (default=0.5)}

\item{poisson_inst}{the poisson ratio for the instantaneous modulus (default=0.5)}

\item{smfl}{the length of the filter for SG smoothing (default=19)}

\item{EqDataPoints}{the number of data points to be included in the measurements of the equilibrium modulus, add 1 to the desired values (default is 201)}
}
\value{
A list containing; folderoutput:A data frame with calculated moduli for all files before and after smoothing, and before and after Hayes correction (1972; based on poisson ration of 0.5);tdfn: A data frame with merged forces and time data extracted from the stress-relaxation section from all files; tdfn_sm: A data frame with merged forces and time data extracted from the stress-relaxation section from all files after smoothing.
}
\description{
Extract the Stress-Relaxation data from MACH-1 files with 3 steps Stress-Relaxation, perpocess the extracted data, and use them to calculate the elastic modulus
}
