% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Find_Contact.R
\name{Extract_Find_Contact}
\alias{Extract_Find_Contact}
\title{Extract_Find_Contact}
\usage{
Extract_Find_Contact(x, y = 11)
}
\arguments{
\item{x}{path to the MACH-1 text file}

\item{y}{the number of lines after the tag "<Find Contact>" (default is 11)}
}
\value{
a data frame with the extracted Find Contact section
}
\description{
Extract the "Find Contact" section from MACH-1 file
}
