% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Single_Stress_Relaxation.R
\name{Extract_Single_Stress_Relaxation}
\alias{Extract_Single_Stress_Relaxation}
\title{Extract_Single_Stress_Relaxation}
\usage{
Extract_Single_Stress_Relaxation(x, start_index, end_index, z = 1)
}
\arguments{
\item{x}{character vectors of data produced by biomomentum MACH-1}

\item{start_index}{the index of the beginning of the data of the Stress-Relaxation cycle ("<Stress Relaxation>" or "<divider>"); the extracted data start after z (the third argument of this function) + this index}

\item{end_index}{the index of the end of the data of the Stress-Relaxation cycle ("<divider>"); the extracted data end just before this index}

\item{z}{the number of the first index to be extracted after the start_index}
}
\value{
a data frame with a single cycle of stress-relaxation data extracted
}
\description{
Extract a Single Stress-Relaxation cycle
}
