% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_SR_Parameters.R
\name{Extract_SR_Parameters}
\alias{Extract_SR_Parameters}
\title{Extract_SR_Parameters}
\usage{
Extract_SR_Parameters(SRmergedfile, Step2Index, Step3Index)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple files}

\item{Step2Index}{the index of the beginning of the second step of the Stress-Relaxation cycles}

\item{Step3Index}{the index of the beginning of the third step of the Stress-Relaxation cycles}
}
\value{
a list of the equilibrium forces (Se1, Se2 & Se3), the peak forces (Sp1, Sp2 & Sp3), and the forces just before the peaks (Sm1, Sm2, & Sm3) of the three steps, respectively.
}
\description{
Extract parameters for calculation of the moduli from extracted Stress-Relaxation data
}
