% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_SR_Parameters2.R
\name{Extract_SR_Parameters2}
\alias{Extract_SR_Parameters2}
\title{Extract_SR_Parameters2}
\usage{
Extract_SR_Parameters2(
  SRmergedfile,
  Step2Index,
  Step3Index,
  Step4Index = 0,
  EqDataPoints = 201
)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple files}

\item{Step2Index}{the index of the beginning of the second step of the Stress-Relaxation cycles}

\item{Step3Index}{the index of the beginning of the third step of the Stress-Relaxation cycles}

\item{Step4Index}{the index of the beginning of the fourth step of the Stress-Relaxation cycles (default is 0 ie ignored)}

\item{EqDataPoints}{the number of data points to be included in the measurements of the equilibrium modulus, add 1 to the desired values (default is 201)}
}
\value{
a list of the equilibrium forces (Se1, Se2, Se3, and optionally Se4), the peak forces (Sp1, Sp2, Sp3, and optionally Sp4), and the forces just before the peaks (Sm1, Sm2, Sm3, and optionally Sm4) of the three/four steps, respectively.
}
\description{
Extract parameters for calculation of the moduli from extracted Stress-Relaxation data (3-4 steps), compared to the 1st version a new option added to adjust the data points to be considered, and allow adding a 4th step
}
