% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Stress_Relaxation.R
\name{Extract_Stress_Relaxation}
\alias{Extract_Stress_Relaxation}
\title{Extract_Stress_Relaxation}
\usage{
Extract_Stress_Relaxation(file, y = 12)
}
\arguments{
\item{file}{path to the MACH-1 text file}

\item{y}{the number of lines after the tag "<Stress Relaxation>" (default is 12)}
}
\value{
a data frame with the extracted 3 stress-relaxation data cycles combined
}
\description{
Extract the Stress-Relaxation cycles
}
