% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Stress_Relaxation_MultiFiles.R
\name{Extract_Stress_Relaxation_MultiFiles}
\alias{Extract_Stress_Relaxation_MultiFiles}
\title{Extract_Stress_Relaxation_MultiFiles}
\usage{
Extract_Stress_Relaxation_MultiFiles(folder_path)
}
\arguments{
\item{folder_path}{path to the folders with MACH-1 text files}
}
\value{
A list containing; tdfn: A data frame with merged forces and time data extracted from the stress-relaxation section from all files; SRDisplacement: A data frame with merged displacement and time data extracted from the stress-relaxation section from all files; StepsIndex: A vector containing the indices of the 3 steps in the stress-relaxation section.
}
\description{
Extract Stress-Relaxation sections from mutliple files in a folder and do essential preprocessing on them
}
