% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Wait.R
\name{Extract_Wait}
\alias{Extract_Wait}
\title{Extract_Wait}
\usage{
Extract_Wait(x, y = 4)
}
\arguments{
\item{x}{path to the MACH-1 text file}

\item{y}{the number of lines after the tag "<Wait>" (default is 4)}
}
\value{
a data frame with the extracted 3 stress-relaxation data cycles combined
}
\description{
Extract the Wait period prior the stress relaxation cycles (if defined in the protocol)
}
