% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hayes_correction.R
\name{Hayes_correction}
\alias{Hayes_correction}
\title{Hayes_correction}
\usage{
Hayes_correction(Modulus, indenterradius, Thicknesses, v = 0.5)
}
\arguments{
\item{Modulus}{the modulus to be corrected}

\item{indenterradius}{indenter radius in mm}

\item{Thicknesses}{the thickness of the sample or list of thicknesses from the 3 step stress relaxation protocol}

\item{v}{Poisson’s ratio (default 0.5; also data provided )}
}
\value{
the corrected Modulus
}
\description{
Correct the modulus according to Hayes el al. (1972), only suitable for materials with Poisson’s ratio = 0.5 such as hydrogels
}
\examples{
Hayes_correction(0.03,1.5,c(2,1.9,1.8),v=0.5)
}
