% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeBaseline.R
\name{NormalizeBaseline}
\alias{NormalizeBaseline}
\title{NormalizeBaseline}
\usage{
NormalizeBaseline(SRmergedfile, dpi = 10)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple MACH-1 files}

\item{dpi}{number of initial datapoint to be considered as the baseline}
}
\value{
a data frame (SRmergedfile_Clean) with all stress data subtracted from the baseline i.e. normalized to initial of Zero
}
\description{
Normalize the baseline of the stress relaxation data to zero
}
