% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmv_abnormal_peaks.R
\name{Rmv_abnormal_peaks}
\alias{Rmv_abnormal_peaks}
\title{Rmv_abnormal_peaks}
\usage{
Rmv_abnormal_peaks(
  SRmergedfile,
  S1i,
  S2i,
  S3i,
  S4i = 0,
  dp = 200,
  threshold = 1.4
)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple MACH-1 files}

\item{S1i}{The index of the 1st step}

\item{S2i}{The index of the 2nd step}

\item{S3i}{The index of the 3rd step}

\item{S4i}{The index of the 4th step (default=0 i.e. just 3 steps)}

\item{dp}{The number of data points to consider around each peak}

\item{threshold}{How much higher is the maximum peak from the second higher stress that makes it abnormal}
}
\value{
a modified data frame (SRmergedfile) with the forces of Stress-Relaxation cleaned from abnormally high peaks
}
\description{
Remove abnormally high peaks around the three/Four stress peaks of the stress relaxation protocol
}
