% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rmv_negative_peaks2.R
\name{Rmv_negative_peaks2}
\alias{Rmv_negative_peaks2}
\title{Rmv_negative_peaks2}
\usage{
Rmv_negative_peaks2(SRmergedfile, S1i, S2i, S3i, S4i = 0)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple MACH-1 files}

\item{S1i}{The index of the 1st step}

\item{S2i}{The index of the 2nd step}

\item{S3i}{The index of the 3rd step}

\item{S4i}{The index of the 4th step (default=0 i.e. just 3 steps)}
}
\value{
a data frame (SRmergedfile_Clean) with the forces of Stress-Relaxation cleaned from negative peaks
}
\description{
Remove negative peaks around the three/Four stress peaks of the stress relaxation protocol
}
