% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_SR.R
\name{smooth_SR}
\alias{smooth_SR}
\title{smooth_SR}
\usage{
smooth_SR(SRmergedfile, smfl = 17)
}
\arguments{
\item{SRmergedfile}{A data frame with merged forces and time data extracted from the stress-relaxation section from multiple files}

\item{smfl}{The length of the filter (default=17)}
}
\value{
a data frame (tdfn_sm) with the forces of Stress-Relaxation smoothed with less noises
}
\description{
Smooth the forces of Stress-Relaxation data by SG filters
}
